[#4] refactor: Rename entities package name from `domain` to `domain.entity`
entity들이 위치한 패키지를 domain 하위의 entity로 변경
