+++ BE/sidedish/src/main/java/team16/sidedish/repository/ProductRepository.java
@@ -2,7 +2,7 @@ package team16.sidedish.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import team16.sidedish.domain.product.Product;
+import team16.sidedish.domain.entity.product.Product;
 
 @Repository
 public interface ProductRepository extends CrudRepository<Product, Long> {
