+++ BE/sidedish/src/main/resources/application.properties
@@ -1,5 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/sidedish
 spring.datasource.username=sidedish
 spring.datasource.password=test
+spring.datasource.initialization-mode=always
 logging.level.sql=debug
 logging.level.team16=debug
